+++ pycoffer/plugins/rsync_cli.py
@@ -9,21 +9,17 @@ import click
 
 from pycoffer import main_lib
 
-# ~ @click.group(help='Manage passwords in coffer.')
-# ~ @click.command(help='Manage passwords in coffer.')
-# ~ def password():
-    # ~ pass
-
 @click.group()
 def cli():
     pass
 
-@cli.command(help='Rsync directory with keys of coffer (to do).')
+@cli.command(help='Rsync directory with keys of coffer using mtime (to do).')
 @main_lib.opt_configuration
 @main_lib.opt_coffer
 @click.option('-s', "--source", help='The source directory.')
 @click.option('-t', "--target", help='The target directory in coffer.')
-def rsync(conf, coffer, source, target):
+@click.option("--dry", is_flag=True, show_default=True, default=False, help="Don't do anything, only show files that will be updated.")
+def rsync(conf, coffer, source, target, dry):
     with main_lib.open_coffer(conf, coffer, 'a') as ff:
         with ff.plugin('rsync') as plg:
-            plg.rsync(source, target)
+            plg.rsync(source, target, dry)