+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 
 - Move coffer definitions in coffers module
+- Finish rsync plugin
 
 ### Removed